+++ forui/lib/src/theme/theme_data.dart
@@ -148,7 +148,7 @@ final class FThemeData with Diagnosticable {
       tooltipStyle: FTooltipStyle.inherit(colorScheme: colorScheme, typography: typography, style: style),
       scaffoldStyle: FScaffoldStyle.inherit(colorScheme: colorScheme, style: style),
       dividerStyles: FDividerStyles.inherit(colorScheme: colorScheme, style: style),
-      switchStyle: FSwitchStyle.inherit(colorScheme: colorScheme),
+      switchStyle: FSwitchStyle.inherit(colorScheme: colorScheme, style: style),
     );
   }